requirements.txt
sqlalchemy==1.3.20
black==20.8b1
flake8==3.8.4
psycopg2-binary==2.8.6
python-dotenv==0.14.0